+++ CHANGELOG.md
@@ -15,5 +15,8 @@
 
   While the prior comma separated default was convenient for the two cases above, it interfered with `UPDATE models SET tags = ${tags}`. It may also have generated invalid SQL if you didn't expect some value to be an array. It wouldn't have caused a security issue, but invalid SQL wasn't great either.
 
+- Adds support for calling `Sql.prototype.toString` with other placeholder formats.
+- Fixes PostgreSQL support by using `$` placeholders, not `?`s.
+
 ## 1.0.0 (Apr 27, 2019)
 - Things could sqlate quickly.